requirements.txt
# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Computer Vision and Image Processing
opencv-contrib-python==4.9.0.80
numpy==1.26.0
Pillow==10.3.0

# Machine Learning (compatible with Python 3.13)
scikit-learn==1.4.0

# Image encoding/decoding
imageio==2.33.0

# For better error handling and logging
Werkzeug==3.0.1

# For JSON handling and utilities
six==1.16.0

# Memory optimization
psutil==5.9.6

# Development and testing (optional)
pytest==7.4.4
pytest-flask==1.3.0

# Production server (optional)
gunicorn==21.2.0
